+++ spec/configuration_spec.rb
@@ -6,8 +6,25 @@ describe ENVied::Configuration do
   it { is_expected.to respond_to :defaults_enabled? }
 
   describe '#variable' do
-    it 'results in an added variable' do
+    def with_envfile(&block)
+      @config = described_class.new(&block)
+    end
+    attr_reader :config
 
+    it 'results in an added variable' do
+      with_envfile do
+        variable :foo, :boolean
+      end
+
+      expect(config.variables).to include ENVied::Variable.new(:foo, :boolean)
+    end
+
+    it 'sets string as type when no type is given' do
+      with_envfile do
+        variable :bar, default: 'bar'
+      end
+
+      expect(config.variables).to include ENVied::Variable.new(:bar, :string, default: 'bar')
     end
   end